+++ fastimer/stop_fast.py
@@ -5,25 +5,12 @@ import datetime
 # noinspection PyPackageRequirements
 from consolemenu import PromptUtils
 
-from .fasts_file import read_fasts, write_fasts
-from .utils import get_active_fast
+from .fasts_file import write_fasts
 
 
-def stop_fast(prompt: PromptUtils) -> None:
+def stop_fast(prompt: PromptUtils, fasts: list, fast: dict) -> None:
 
-    fasts = read_fasts()
-    fast = get_active_fast(fasts)
+    if prompt.prompt_for_yes_or_no("Do you want to end your ongoing fast?"):
 
-    if fast is not None:
-
-        if prompt.prompt_for_yes_or_no("Do you want to end your ongoing fast?"):
-
-            fasts[-1]["stopped"] = datetime.datetime.now()
-            write_fasts(fasts)
-
-    else:
-
-        print("No current fast to end.")
-        print()
-
-        prompt.enter_to_continue()
+        fasts[-1]["stopped"] = datetime.datetime.now()
+        write_fasts(fasts)